Frame skip implemented
Now when multiple game update messages are received within a single frame, the client will only use the most recent one for the frame, removing unnecessary updates from the loop.

Also stumbled onto a nasty issue with C++ regarding what pointers actually look at in memory. See Main.cpp lines 69-72 comments for a rant explaining it.
 ^ Long story short, the issue was fixed by storing update message data as a char[] instead of a char* to avoid pointing at a memory location that constantly changes.

This was meant to take about 30 minutes to implement, but that issue turned it into a 12 hour long nightmare.

Addon to last commit: Another optimisation I forgot to mention is that I programmed some parts way too defensively, with a lot of checks being done that could be ignored with just a little bit of trust in the error case not happening. So later on some of these safety checks will be removed for the sake of performance.

*Note, comments may contain bad language. Not much, but there would have been a lot more if this wasn't a university project.*