Sync time implemented & start of entity replication
A time reference can now be passed into messages to allow clients to handle them while compensating for ping and other delays.
^ Both the client and server had functions that return OS time in milliseconds and they work together to make sure everything is in the right place.
^ Right now the client is actually a little bit ahead when using this because when spawning projectiles, the time on server side is recorded as the projectile is created, a while before it's first tick update. Fixing this should be easy enough, I could even just add a manual offset to not overcomplicate things.

The general structure for handling attacks and projectiles is being made. All 3 boss attacks can be replicated to the client now and they generally run pretty well aligned to the server at 60hz. There is support for gravity, but properly implementing player attacks will be more difficult as they disappear on collision, after several hits, or on a timer, meaning that I might need to either update their death through the server or simulate collisions for those projectiles on clientside.

TODO:
 - A proper fail state needs to be implemented on the server.
 - Text needs to be added to UI
 - UI needs to be implemented on client.
 - Boss/NPCs need to be replicated
 - Player attacks need to be replicated
 - Continue working on projectiles to make sure they're as good as they can be.
 - Other optimisations