Texture optimisation & data restructure
All textures used in the game are loaded into memory at launch. Entities that need one of those textures can just copy the relevant pointer and use it for rendering.

This thing was meant to take 5 minutes to implement, but this is C++ so of course it took 4 days instead. I had to find out about how C++ handles global and static variables between files and found out that static variables are unique to each file that implements them, so each file's use of "game_data" referenced a different instance of it.
 ^ This concept was actually a major problem a lot earlier in development but I worked around it with the getGameData() function in MyGame.
 ^ I also found out that the global scope used by a class method depends on the file that was the first to call that function, all other calls use the same global scope no matter which file the function was called from.
 ^^ Out of all the C++ quirks that made me want to jump out a window, this one made the least sense to me. I can kind of see how that happens but it's horrible to work with.

Because of the above problem, GameData and other data-like/entity classes were migrated over to a new file "GameData.h" and "GameData.cpp". This restructure allowed me to define game_data as a proper global variable.
 ^ Though when thinking about it, the easiest solution would have been making game_data a variable within the MyGame class as it really doesn't need to be global.
 ^^ If I want to get rid of the global variable later on then I will use that approach instead, switching back won't take long to implement.

On another topic, this change removes textures as a factor of memory usage since they don't get loaded per entity instance anymore, just 6 tiny images existing in memory. That number might go up later but that's not the point.
The only thing that dynamically changes memory now is variables being created and destroyed, and just like the last commit, it does feel like some tiny leak does exist somewhere in the code. I'm confident that all the entities are being cleared correctly in the main sections of the code, but there probably is some part that I missed that's causing the leak; either that or the update message buffer really is the only leak in the entire source code. Either way, memory is going up and sometimes going back down before going up again, so something is not right.

Plans for what to implement are unchanged. Make projectiles better, render the boss, UI and player attacks on client side, and that's more or less it.