Credit
Added references that were used while trying to fix the global scope problem.

Even though this change will be reverted while cleaning up the code before handing it in, I should still credit it as it was an important part of the development process and should be noted down for the report.