UI Start & Fixes
Player health bars and special stat bars now get rendered to the screen. RechargableComponent was effectively brought into the client through the RechargableValue struct, allowing players to store one special value to track depending on class. Fonts were also added so the game can render text for entity names and general info.
 - UI render function is only very simple for now, boss UI is yet to be added and all values don't update just yet.
 ^ Because we don't have nodes and alignments in SDL like we do in JavaFX, the player UI isn't aligned 1:1 with how it is on the server, being offset slightly with multiple players added.

A cool quirk was found where checking unordered_map via array index adds that slot to it's size even if the slot is empty. This makes map.Size() unreliable if I want to keep nullptr checks as the standard getKey() functions crash when the key returns nothing instead of just giving nullptr.

A crash was found when a player exits the game. The player's before them in the list crash but the rest don't - This might be wrong but that was the pattern in testing.
 ^ This was caused by the update message buffer trying to update a player that left the game, because the update message gets delayed until update is ran, there is enough time for the player to get kicked while their message is still in the buffer.
 ^ This was fixed by adding a nullptr check to the referenced player
 ^^ Not every function has a proper nullptr check yet but they will all be added eventually.

Up next is adding boss UI rendering, and replicating health and special stat values to the client.